requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
readability-lxml>=0.8.1
playwright>=1.45.0
tldextract>=5.1.0